requirements.txt
telethon
pandas
amharicNLP
python-dotenv
